+++ CHANGELOG.md
@@ -3,6 +3,7 @@
   - เพิ่ม source HIS: VPM
   - เปลี่ยนตัวแปร Key ให้สอดคล้องกับ MOPH ERP
   - เพิ่ม getVisitForMophAlert
+  - เพิ่มเงื่อนไขใน HOSxP ให้ sum เฉพาะ ward.ward_active=Y
 
 # 3.8.13
 ## Feature